backend/requirements.txt
Flask==2.0.2
Flask-MySQLdb==1.0.1
